Fix packet classes and make the build work

+++ 1.12/src/main/scala/org/tuubes/craft/mc1_12/play/clientbound/BlockBreakAnimationPacket.scala
@@ -20,7 +20,7 @@ final class BlockBreakAnimationPacket(var entityId: Int, var location: Long, var
     val x = location >> 38
     val y = (location >> 26) & 0xfff
     val z = location << 38 >> 38
-    new Vec3i(x, y, z)
+    new Vec3i(x.toInt, y.toInt, z.toInt)
   }	
 }
 object BlockBreakAnimationPacket extends PacketObj[CraftAttach, BlockBreakAnimationPacket] {
