+++ 1.12/src/main/scala/org/tuubes/craft/mc1_12/play/clientbound/BlockChangePacket.scala
@@ -19,7 +19,7 @@ final class BlockChangePacket(var location: Long, var blockId: Int) extends Pack
     val x = location >> 38
     val y = (location >> 26) & 0xfff
     val z = location << 38 >> 38
-    new Vec3i(x, y, z)
+    new Vec3i(x.toInt, y.toInt, z.toInt)
   }	
 }
 object BlockChangePacket extends PacketObj[CraftAttach, BlockChangePacket] {
