+++ 1.12/src/main/scala/org/tuubes/craft/mc1_12/play/clientbound/UseBedPacket.scala
@@ -19,7 +19,7 @@ final class UseBedPacket(var entityId: Int, var location: Long) extends Packet[C
     val x = location >> 38
     val y = (location >> 26) & 0xfff
     val z = location << 38 >> 38
-    new Vec3i(x, y, z)
+    new Vec3i(x.toInt, y.toInt, z.toInt)
   }	
 }
 object UseBedPacket extends PacketObj[CraftAttach, UseBedPacket] {
