+++ 1.12/src/main/scala/org/tuubes/craft/mc1_12/play/serverbound/PlayerBlockPlacementPacket.scala
@@ -23,7 +23,7 @@ final class PlayerBlockPlacementPacket(var location: Long, var face: Int, var ha
     val x = location >> 38
     val y = (location >> 26) & 0xfff
     val z = location << 38 >> 38
-    new Vec3i(x, y, z)
+    new Vec3i(x.toInt, y.toInt, z.toInt)
   }	
 }
 object PlayerBlockPlacementPacket extends PacketObj[CraftAttach, PlayerBlockPlacementPacket] {
