+++ 1.12/src/main/scala/org/tuubes/craft/mc1_12/play/serverbound/UpdateSignPacket.scala
@@ -23,7 +23,7 @@ final class UpdateSignPacket(var location: Long, var line1: String, var line2: S
     val x = location >> 38
     val y = (location >> 26) & 0xfff
     val z = location << 38 >> 38
-    new Vec3i(x, y, z)
+    new Vec3i(x.toInt, y.toInt, z.toInt)
   }	
 }
 object UpdateSignPacket extends PacketObj[CraftAttach, UpdateSignPacket] {
